exchanged convolution distribution
