+++ genomicsurveillance/__init__.py
@@ -14,13 +14,10 @@ from .models import (
 )
 from .utils import (
     KnotList,
-    aggregate_tensor,
-    alias_lineages,
     create_date_list,
     create_spline_basis,
     epiestim_discretise_serial_interval,
     epiestim_R,
-    merge_lineages,
     preprocess_lineage_tensor,
 )
 
@@ -30,10 +27,7 @@ __all__ = [
     "get_specimen",
     "create_spline_basis",
     "create_date_list",
-    "merge_lineages",
-    "alias_lineages",
     "preprocess_lineage_tensor",
-    "aggregate_tensor",
     "epiestim_R",
     "epiestim_discretise_serial_interval",
     "MultiLineage",
